requirements.txt
pygsheets
google-api-python-client
google-auth-oauthlib
google-auth
python-dotenv